refactor: remove baseCreateViewModel and enhance the home screen and move the logic to homeViewModel
