+++ app/src/main/java/com/london/tudee/di/ViewModelModule.kt
@@ -10,7 +10,6 @@ import com.london.tudee.presentation.screens.categories.EditCategoryScreenViewMo
 import com.london.tudee.presentation.screens.categories.crud.CreateCategoryScreenViewModel
 import com.london.tudee.presentation.screens.categories.crud.DeleteCategoryScreenViewModel
 import com.london.tudee.presentation.screens.onboarding.OnBoardingViewModel
-import com.london.tudee.presentation.screens.task.add_edit_task_bottom_sheet.AddOrEditTaskViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
@@ -25,5 +24,4 @@ val viewModelModule = module {
     viewModelOf(::CreateCategoryScreenViewModel)
     viewModelOf(::DeleteCategoryScreenViewModel)
     viewModelOf(::OnBoardingViewModel)
-    viewModelOf(::AddOrEditTaskViewModel)
 }
